refactor(SlidingPiece) : 중복 코드 제거
