+++ src/main/java/chess/domain/chesspiece/slidingPiece/King.java
@@ -1,5 +1,8 @@
-package chess.domain.chesspiece;
+package chess.domain.chesspiece.slidingPiece;
 
+import chess.domain.chesspiece.Piece;
+import chess.domain.chesspiece.Role;
+import chess.domain.chesspiece.Team;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
 
@@ -9,25 +12,10 @@ import java.util.List;
 
 import static chess.domain.chesspiece.Role.*;
 
-public class King extends Piece {
+public class King extends SlidingPiece {
 
     public King(Team team) {
         super(team);
-    }
-
-    @Override
-    public List<Position> getRoute(Position source, Position target) {
-        List<Position> route = new ArrayList<>();
-        validateMovingRule(source, target);
-
-        Direction direction = Direction.findDirection(source, target);
-        Position movingPosition = direction.move(source);
-
-        while (!movingPosition.equals(target)) {
-            route.add(movingPosition);
-            movingPosition = direction.move(movingPosition);
-        }
-        return Collections.unmodifiableList(route);
     }
 
     @Override
@@ -45,10 +33,5 @@ public class King extends Piece {
             return WHITE_KING;
         }
         return BLACK_KING;
-    }
-
-    @Override
-    public boolean isPawn() {
-        return false;
     }
 }
