+++ src/main/java/chess/domain/chesspiece/slidingPiece/SlidingPiece.java
@@ -1,18 +1,17 @@
-package chess.domain.chesspiece;
+package chess.domain.chesspiece.slidingPiece;
 
+import chess.domain.chesspiece.Piece;
+import chess.domain.chesspiece.Role;
+import chess.domain.chesspiece.Team;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static chess.domain.chesspiece.Role.BLACK_BISHOP;
-import static chess.domain.chesspiece.Role.WHITE_BISHOP;
+public abstract class SlidingPiece extends Piece {
 
-public class Bishop extends Piece {
-
-    public Bishop(Team team) {
+    public SlidingPiece(Team team) {
         super(team);
     }
 
@@ -29,23 +28,6 @@ public class Bishop extends Piece {
             movingPosition = direction.move(movingPosition);
         }
         return Collections.unmodifiableList(route);
-    }
-
-    @Override
-    protected void validateMovingRule(Position source, Position target) {
-        int fileDistance = source.calculateFileDistance(target);
-        int columnDistance = source.calculateRankDistance(target);
-        if (fileDistance != columnDistance) {
-            throw new IllegalArgumentException("이동할 수 없습니다.");
-        }
-    }
-
-    @Override
-    public Role getRole() {
-        if (getTeam().isWhite()) {
-            return WHITE_BISHOP;
-        }
-        return BLACK_BISHOP;
     }
 
     @Override
